:truck: Rename and move files to fit MC of MVC pattern
Models and controllers have been moved into separate folders, controllers following single-case naming convention and models following plural-case
